SoA indexing overflow fix
fixed SoA indexing overflow by changing the indexing policy from unsigned short (16bits, 0-65'535) to unsigned int (32bits, 0-4'294'967'295) to support large strides

+++ include/TFEL/Math/Array/StridedCoalescedViewsArray.hxx
@@ -49,9 +49,7 @@ namespace tfel::math::internals {
    * \tparam ViewIndexingPolicyType: indexing policy used by the view
    */
   template <typename MappedType, typename ViewIndexingPolicyType>
-  struct StridedCoalescedViewsArrayBase<true,
-                                        MappedType,
-                                        ViewIndexingPolicyType> {
+  struct StridedCoalescedViewsArrayBase<true, MappedType, ViewIndexingPolicyType> {
     static_assert(tfel::math::isScalar<MappedType>());
     //! \brief type of views
     using view_type = MappedType&;
@@ -81,7 +79,7 @@ namespace tfel::math::internals {
   struct StridedCoalescedViewsArrayDefaultViewIndexingPolicy<true, MappedType> {
     static_assert(tfel::math::isScalar<MappedType>());
     //! \brief default indexing type
-    using type = tfel::math::ScalarIndexingPolicy<unsigned short>;
+    using type = tfel::math::ScalarIndexingPolicy<unsigned int>;
   };  // end of StridedCoalescedViewsArrayDefaultViewIndexingPolicy
 
 }  // end of namespace tfel::math::internals
@@ -97,10 +95,10 @@ namespace tfel::math {
 
   //! \brief the pointer type used to build the view
   template <typename MappedType>
-  using StridedCoalescedViewsArrayDataPointerType = std::conditional_t<
-      std::is_const_v<MappedType>,
-      const StridedCoalescedViewsArrayNumericType<MappedType>*,
-      StridedCoalescedViewsArrayNumericType<MappedType>*>;
+  using StridedCoalescedViewsArrayDataPointerType =
+      std::conditional_t<std::is_const_v<MappedType>,
+                         const StridedCoalescedViewsArrayNumericType<MappedType>*,
+                         StridedCoalescedViewsArrayNumericType<MappedType>*>;
 
   /*!
    * \brief a structure used to represent an array of strided coalesced views.
@@ -111,16 +109,17 @@ namespace tfel::math {
    */
   template <typename MappedType,
             typename MemoryIndexingPolicyType,
-            typename ViewIndexingPolicyType = typename tfel::math::internals::
-                StridedCoalescedViewsArrayDefaultViewIndexingPolicy<
-                    isScalar<MappedType>(),
-                    MappedType>::type>
+            typename ViewIndexingPolicyType =
+                typename tfel::math::internals::
+                    StridedCoalescedViewsArrayDefaultViewIndexingPolicy<
+                        isScalar<MappedType>(),
+                        MappedType>::type>
   struct StridedCoalescedViewsArray
       : MemoryIndexingPolicyType,
-        tfel::math::internals::StridedCoalescedViewsArrayBase<
-            isScalar<MappedType>(),
-            MappedType,
-            ViewIndexingPolicyType> {
+        tfel::math::internals::
+            StridedCoalescedViewsArrayBase<isScalar<MappedType>(),
+                                           MappedType,
+                                           ViewIndexingPolicyType> {
     //! \brief boolean stating if views to scalars are considered
     static constexpr bool is_scalar = isScalar<MappedType>();
     //! \brief type of the pointer to the underlying data
@@ -164,17 +163,15 @@ namespace tfel::math {
       static_assert(!MemoryIndexingPolicyType::hasFixedSizes, "invalid call");
     }
     //! \brief copy constructor
-    constexpr StridedCoalescedViewsArray(
-        const StridedCoalescedViewsArray&) noexcept = default;
+    constexpr StridedCoalescedViewsArray(const StridedCoalescedViewsArray&) noexcept = default;
     //! \brief move constructor
-    constexpr StridedCoalescedViewsArray(
-        StridedCoalescedViewsArray&&) noexcept = default;
+    constexpr StridedCoalescedViewsArray(StridedCoalescedViewsArray&&) noexcept = default;
     //! \brief a simple alias
     using StridedCoalescedViewsArrayBase =
-        tfel::math::internals::StridedCoalescedViewsArrayBase<
-            isScalar<MappedType>(),
-            MappedType,
-            ViewIndexingPolicyType>;
+        tfel::math::internals::
+            StridedCoalescedViewsArrayBase<isScalar<MappedType>(),
+                                           MappedType,
+                                           ViewIndexingPolicyType>;
     //! \brief access operator
     TFEL_HOST_DEVICE constexpr typename StridedCoalescedViewsArray::view_type
     operator[](const typename MemoryIndexingPolicyType::size_type i) noexcept {
@@ -190,8 +187,8 @@ namespace tfel::math {
     //! \brief access operator
     TFEL_HOST_DEVICE constexpr
         typename StridedCoalescedViewsArray::const_view_type
-        operator[](const typename MemoryIndexingPolicyType::size_type i)
-            const noexcept {
+    operator[](const typename MemoryIndexingPolicyType::size_type i)
+        const noexcept {
       static_assert(MemoryIndexingPolicyType::arity == 1u, "invalid call");
       static_assert(ViewIndexingPolicyType::hasFixedSizes, "invalid call");
       auto* const d = this->data_pointer + i;
@@ -218,7 +215,7 @@ namespace tfel::math {
     template <typename... Indices>
     TFEL_HOST_DEVICE constexpr
         typename StridedCoalescedViewsArray::const_view_type
-        operator()(const Indices... i) const noexcept {
+    operator()(const Indices... i) const noexcept {
       static_assert(ViewIndexingPolicyType::hasFixedSizes, "invalid call");
       const auto* const d = this->data_pointer + this->getIndex(i...);
       if constexpr (is_scalar) {
@@ -243,13 +240,15 @@ namespace tfel::math {
         return p.size(i);
       };
       if constexpr (MemoryIndexingPolicyType::unRollLoop) {
-        IterateOverMultipleIndices<
-            typename StridedCoalescedViewsArray::size_type, 0,
-            MemoryIndexingPolicyType::arity, true>::exe(f, g);
+        IterateOverMultipleIndices<typename StridedCoalescedViewsArray::size_type,
+                                   0,
+                                   MemoryIndexingPolicyType::arity,
+                                   true>::exe(f, g);
       } else {
-        IterateOverMultipleIndices<
-            typename StridedCoalescedViewsArray::size_type, 0,
-            MemoryIndexingPolicyType::arity, false>::exe(f, g);
+        IterateOverMultipleIndices<typename StridedCoalescedViewsArray::size_type,
+                                   0,
+                                   MemoryIndexingPolicyType::arity,
+                                   false>::exe(f, g);
       }
       return *this;
     }  // end of operator*=
@@ -271,17 +270,19 @@ namespace tfel::math {
     const data_pointer_type data_pointer;
   };  // end of struct StridedCoalescedViewsArray
 
-  template <typename MappedType, typename SizeType, SizeType N>
+  template <typename MappedType,
+            typename SizeType,
+            SizeType N>
   struct StridedCoalescedViewsFixedSizeVectorIndexingPolicy
       : FixedSizeVectorIndexingPolicy<SizeType, N, 1> {
   };  // end of struct StridedCoalescedViewsFixedSizeVectorIndexingPolicy
 
-  template <typename MappedType, typename SizeType, SizeType N>
+  template <typename MappedType,
+            typename SizeType,
+            SizeType N>
   using StridedCoalescedViewsFixedSizeVector = StridedCoalescedViewsArray<
       MappedType,
-      StridedCoalescedViewsFixedSizeVectorIndexingPolicy<MappedType,
-                                                         SizeType,
-                                                         N>>;
+      StridedCoalescedViewsFixedSizeVectorIndexingPolicy<MappedType, SizeType, N>>;
 
   template <typename ArrayType>
   constexpr auto map_strided_array(
@@ -294,8 +295,8 @@ namespace tfel::math {
     static_assert(IndexingPolicy::arity == 1u);
     constexpr auto N = IndexingPolicy().size(0);
     if constexpr (std::is_const_v<ArrayType>) {
-      return StridedCoalescedViewsFixedSizeVector<const MappedType, size_type,
-                                                  N>(p);
+      return StridedCoalescedViewsFixedSizeVector<const MappedType, size_type, N>(
+          p);
     } else {
       return StridedCoalescedViewsFixedSizeVector<MappedType, size_type, N>(p);
     }
